feat(ch-app): Refactor logging-api to use a service as well
- Separated rocket-handler and LoggingService method, so that a http server replacement becomes easier and the architecture becomes more modular.
- Added ports module, where the rocket-handler code will be moved in the next commits
- Added forbid(unsafe_code) in every crate
