+++ clearing-house-app/logging-service/src/ports/mod.rs
@@ -1,0 +1,5 @@
+//! # Ports
+//!
+//! This module contains the ports of the logging service. Ports are used to communicate with other
+//! services. In this case, the logging service implements REST-API endpoints to provide access to
+//! the logging service.
