+++ clearing-house-app/logging-service/src/services/mod.rs
@@ -1,2 +1,9 @@
+//! # Services
+//!
+//! This module contains the Application Services that are used by the API Controllers. It is
+//! responsible for the business logic of the application. The services are used by the API
+//! Controllers to handle the requests and responses.
+//! 
 mod keyring_service;
 mod document_service;
+pub(crate) mod logging_service;
